deleted unwanted files
