--- BackEnd/Aswin/A2.csv
@@ -1,76 +1,0 @@
-College id,Student Name,Contact no
-TVE-S4_CS2-1,ABHAY O,+91 8547960839
-TVE-S4_CS2-2,ABHINAND J NAMBIAR,+91 8921347402
-TVE-S4_CS2-3,ABHIRAM S MENON,+91 9778138706
-TVE-S4_CS2-4,ADARSH MENON,+91 9061676796
-TVE-S4_CS2-5,AISWARYA M S,+91 8590199574
-TVE-S4_CS2-6,AJITH JOSE,+91 7736957108
-TVE-S4_CS2-7,AKSSHAY MATHEW P,+91 9544049564
-TVE-S4_CS2-8,ALFEEN SADIQ,+91 7025734481
-TVE-S4_CS2-9,ANAMIKA S,+91 9895537625
-TVE-S4_CS2-10,ANANTHAPADMANABHAN V,+91 8281479453
-TVE-S4_CS2-11,ANIRUDH A V,+91 9496960369
-TVE-S4_CS2-12,ANJALI AJU,+91 7592816941
-TVE-S4_CS2-13,ARJUN S RAJ,+91 7592906780
-TVE-S4_CS2-14,A R ASHIQ,+91 9174873430
-TVE-S4_CS2-15,ASHWATHY A,+91 8157940237
-TVE-S4_CS2-16,ASHWIN JACOB GIGO,+91 8921838679
-TVE-S4_CS2-17,ASWANTH R,+91 7012172392
-TVE-S4_CS2-18,ASWIN A,+91 7034634277
-TVE-S4_CS2-19,ASWIN C,+91 9605159952
-TVE-S4_CS2-20,ATHIRA B,+91 9496895715
-TVE-S4_CS2-21,BINCE BABU,+91 8301007670
-TVE-S4_CS2-22,DEMEL C JOHN,+91 8281272471
-TVE-S4_CS2-23,DON JOSHI N,+91 9496164302
-TVE-S4_CS2-24,GOKUL G MENON,+91 9567551970
-TVE-S4_CS2-25,GOVIND P,+91 9947412343
-TVE-S4_CS2-26,HAREESH MOHAN S,+91 9961197039
-TVE-S4_CS2-27,HRIDYA S ANIL,+91 8304064833
-TVE-S4_CS2-28,JERRY PAUL ALEX,+91 9496933257
-TVE-S4_CS2-29,JESVIN JOSE,+91 8304857182
-TVE-S4_CS2-30,KALYAN DATH,+91 9074795860
-TVE-S4_CS2-31,KEERTHI ANIL NAIR,+91 9544270166
-TVE-S4_CS2-32,MEERA SREEKUMAR,+91 9188449599
-TVE-S4_CS2-33,MITHUN JOBY,+91 9747358541
-TVE-S4_CS2-34,MOHAMMED SABITH C,+91 8943108040
-TVE-S4_CS2-35,MUHAMMED AMAL SHAN K T,+91 8113002501
-TVE-S4_CS2-36,NANDANA S,+91 8086786440
-TVE-S4_CS2-37,NAVIN VARGHESE GEORGE,+91 9847756420
-TVE-S4_CS2-38,NIRUPAMA T S,+91 9895304296
-TVE-S4_CS2-39,NIVED NARAYANAN,+91 9447300552
-TVE-S4_CS2-40,PARVATHI BAI M,+91 7736613891
-TVE-S4_CS2-41,PRAJWAL G NAIK,+91 6361385262
-TVE-S4_CS2-42,PREJITH P,+91 9778028748
-TVE-S4_CS2-43,RAHUL C MADHAVAN,+91 7510447992
-TVE-S4_CS2-44,RAYAN M N,+91 9946018912
-TVE-S4_CS2-45,RITHWIK BAGE,+91 7063916269
-TVE-S4_CS2-46,ROSHAN MAJO,+91 6238554182
-TVE-S4_CS2-47,SARATH RAGHU,+91 8078856069
-TVE-S4_CS2-48,SHIV DEV SAI RAJ,+91 9496749717
-TVE-S4_CS2-49,SIDDHARTH SHANKAR BYLOPPILLY,+91 8588817622
-TVE-S4_CS2-50,SNEHA SANI,+91 8086095935
-TVE-S4_CS2-51,SOURAV K,+91 9496675748
-TVE-S4_CS2-52,SREEHARI C,+91 9946295790
-TVE-S4_CS2-53,STIJO STEPHEN,+91 7902788412
-TVE-S4_CS2-54,SURAJ KUMAR JAYAKUMAR,+91 9048906799
-TVE-S4_CS2-55,UDITH SUDHAKAR,+91 9605171570
-TVE-S4_CS2-56,VAISHNAV M,+91 8943699619
-TVE-S4_CS2-57,VIVEK P CHANDRAN,+91 9645898941
-TVE-S4_CS2-58,CHRISTIN SIBI,+91 9846656129
-TVE-S4_CS2-59,DEVIKA S KUMAR,+91 9495640802
-TVE-S4_CS2-60,DON JOSE MATHEW,+91 6282568769
-TVE-S4_CS2-61,HARITHA DAS,+91 8086478576
-TVE-S4_CS2-62,SANDRA E,+91 6282584148
-TVE-S4_CS2-63,SREEDEV C V,+91 7736971193
-TVE-S4_CS2-64,THOMAS V VARGHESE,+91 8075581672
-TVE-S4_CS2-65,ALLEN BENNY MATHEW,+91 9074286056
-TVE-S4_CS2-66,ARJUN VIJAYAKUMAR,+91 9072194621
-TVE-S4_CS2-67,ADARSH D,+91 9496641497
-TVE-S4_CS2-68,ARYA THAMPI,+91 9072194621
-TVE-S4_CS2-69,MOHAMMED BAZED ALI,+91 9946520134
-TVE-S4_CS2-70,NIRANJANA M S,+91 6282572815
-TVE-S4_CS2-71,ADARSH D,+91 9496641497
-TVE-S4_CS2-72,AMRUTHA P,+91 8075484215
-TVE-S4_CS2-73,THOUKHEER IBNU ALI ASHRAF A,+91 8139841294
-TVE-S4_CS2-74,ASWIN K,+91 7994627349
-TVE-S4_CS2-75,JACOB DEEPU,+91 8943234578
