--- BackEnd/haritha/studentsdatatarget.csv
@@ -1,74 +1,0 @@
-Roll No.,Name,Phone Number,Class
-1,ABHAY O,8547960839,S4_CS2
-2,ABHINAND J NAMBIAR,8921347402,S4_CS2
-3,ABHIRAM S MENON,9778138706,S4_CS2
-4,ADARSH MENON,9061676796,S4_CS2
-5,AISWARYA M S,8590199574,S4_CS2
-6,AJITH JOSE,7736957108,S4_CS2
-7,AKSSHAY MATHEW P,9544049564,S4_CS2
-8,ALFEEN SADIQ,7025734481,S4_CS2
-9,ANAMIKA S RAJ,9895537625,S4_CS2
-10,ANANTHAPADMANABHAN V,8281479453,S4_CS2
-31,KEERTHI ANIL NAIR,9544270166,S4_CS2
-32,MEERA SREEKUMAR,9188449599,S4_CS2
-33,MITHUN JOBY,9747358541,S4_CS2
-34,MOHAMMED SABITH C,8943108040,S4_CS2
-35,MUHAMMED AMAL SHAN K T P,8113002501,S4_CS2
-36,NANDANA S,8086786440,S4_CS2
-37,NAVIN VARGHESE GEORGE,9847756420,S4_CS2
-38,NIRUPAMA T S,9895304296,S4_CS2
-39,NIVED NARAYANAN,9447300552,S4_CS2
-40,PARVATHI BAI M,7736613891,S4_CS2
-41,PRAJWAL G NAIK,6361385262,S4_CS2
-42,PREJITH P,9778028748,S4_CS2
-43,RAHUL C MADHAVAN,7510447992,S4_CS2
-44,RAYAN M N,9946018912,S4_CS2
-45,RITHWIK BAGE,7063916269,S4_CS2
-46,ROSHAN MAJO,6238554182,S4_CS2
-47,SARATH RAGHU,8078856069,S4_CS2
-48,SHIV DEV SAI RAJ,9496749717,S4_CS2
-49,SIDDHARTH SHANKAR BYLOPPILLY,8588817622,S4_CS2
-50,SNEHA SANI,8086095935,S4_CS2
-51,SOURAV K,9496675748,S4_CS2
-52,SREEHARI C,9946295790,S4_CS2
-53,STIJO STEPHEN,7902788412,S4_CS2
-54,SURAJ KUMAR JAYAKUMAR,9048906799,S4_CS2
-55,UDITH SUDHAKAR,9605171570,S4_CS2
-56,VAISHNAV M,8943699619,S4_CS2
-57,VIVEK P CHANDRAN,9645898941,S4_CS2
-58,CHRISTIN SIBI,9846656129,S4_CS2
-59,DEVIKA S KUMAR,9495640802,S4_CS2
-60,DON JOSE MATHEW,6282568769,S4_CS
-11,ANIRUDH A V,9496960369,S4_CS2
-12,ANJALI AJU,7592816941,S4_CS2
-13,ARJUN S,7592906780,S4_CS2
-14,A R ASHIQ,9074873430,S4_CS2
-15,ASHWATHY A,8157940237,S4_CS2
-17,ASWANTH R,7012172392,S4_CS2
-18,ASWIN A,7034634277,S4_CS2
-19,ASWIN C,9605159952,S4_CS2
-20,ATHIRA B,9496895715,S4_CS2
-21,BINCE BABU,8301007670,S4_CS2
-22,DEMEL C JOHN,8281272471,S4_CS2
-23,DON JOSHI N,9496164302,S4_CS2
-24,GOKUL G MENON,9567551970,S4_CS2
-25,GOVIND P,9947412343,S4_CS2
-26,HAREESH MOHAN S,9961197039,S4_CS2
-27,HRIDYA S ANIL,8304064833,S4_CS2
-28,JERRY PAUL ALEX,9496933257,S4_CS2
-29,JESVIN JOSE,8304857182,S4_CS25
-30,KALYAN DATH,9074795860,S4_CS2
-61,HARITHA DAS,8086478576,S4_CS2
-62,SANDRA E,6282584148,S4_CS2
-63,SREEDEV C V,7736971193,S4_CS2
-64,THOMAS V VARGHESE,8075581672,S4_CS2
-65,ALLEN BENNY MATHEW,9074286056,S4_CS2
-66,ARJUN VIJAYAKUMAR,9496641497,S4_CS2
-67,MIDHUN MURALEEDHARAN,9656612087,S4_CS2
-68,ARYA THAMPI,9072194621,S4_CS2
-69,MOHAMMED BAZED ALI,9946520134,S4_CS2
-70,NIRANJANA M S,6282572815,S4_CS2
-71,ADARSH D,9746418043,S4_CS2
-72,AMRUTHA P,8075484215,S4_CS2
-73,THOUKHEER IBNU ALI ASHRAF A,8139841294,S4_CS2
-75,JACOB DEEPU,8943234578,S4_CS2
