--- BackEnd/sampleModels.py
@@ -1,9 +1,0 @@
-import pandas as pd
-
-sampledata1 = pd.DataFrame({
-    'Name': ['Alice', 'Bob', 'Charlie'],
-    'Age': [25, 30, 22],
-    'Location': ['New York', 'San Francisco', 'Los Angeles']
-})
-
-
